Fix #32
Refactor LICTest file, so that each LIC has its own test file
Add tests for LIC#0 and LIC#1
